chore(implement review feedback) update delete functionality

+++ server/controllers/RequestsController.js
@@ -93,7 +93,12 @@ class RequestsController {
     if (status === 'approved') {
       return res.status(400).json({ error: { message: 'Requests being worked on cannot be deleted' } });
     }
-    requestDB.deleteRequest(id)
+    if (status === 'pending') {
+      requestDB.deleteRequest(id)
+        .then(() => res.status(200).json({ success: { message: 'Request has been deleted' } }))
+        .catch(() => res.status(500).json(errors.error500));
+    }
+    requestDB.deleteRequestWithPersistance(id)
       .then(() => res.status(200).json({ success: { message: 'Request has been deleted' } }))
       .catch(() => res.status(500).json(errors.error500));
   }
